requirements.txt
alembic==1.11.2
Flask==2.1.3
Flask_SQLAlchemy==2.4.1
marshmallow==3.20.1
PyJWT==1.7.1
python_dateutil==2.8.2
Requests==2.31.0
SQLAlchemy==2.0.19
Werkzeug==2.0.3
